# apps/api-gateway/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/accounts/{id}": {
      "get": {
        "operationId": "getAccountById",
        "summary": "Get an user account by account id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
...
    "/api/students": {
      "get": {
        "operationId": "indexStudents",
        "parameters": [
          {
            "name": "hasSupervisor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "supervisorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Students"
        ]
      }
    },
    "/api/students/create": {
      "post": {
        "operationId": "createStudent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Students"
        ]
      }
    },
    "/api/students/accounts/{accountId}": {
      "get": {
        "operationId": "getStudentByAccountId",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Students"
        ]
      }
    },
    "/api/supervisors": {
      "get": {
        "operationId": "indexSupervisors",
        "parameters": [
          {
            "name": "fieldId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxCapacity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minCapacity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Supervisors"
        ]
      }
    },
    "/api/supervisors/create": {
      "post": {
        "operationId": "createSupervisor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupervisorBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Supervisors"
        ]
      }
    },
    "/api/supervisors/accounts/{accountId}": {
      "get": {
        "operationId": "getSupervisorByAccountId",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Supervisors"
        ]
      }
    },
    "/api/supervisors/update/{id}": {
      "put": {
        "operationId": "updateSupervisor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupervisorBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Supervisors"
        ]
      }
    }
  },
  "info": {
    "title": "SV Connect API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "AccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "email",
          "emailVerified",
          "role"
        ]
      },
...
      "ConnectSpecializationBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ConnectSupervisorBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "ConnectAccountBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
...
      "ConnectFieldBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "CreateStudentBody": {
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/ConnectAccountBody"
          },
          "supervisor": {
            "$ref": "#/components/schemas/ConnectSupervisorBody"
          }
        },
        "required": [
          "account"
        ]
      },
      "CreateSupervisorBody": {
        "type": "object",
        "properties": {
          "capacity": {
            "type": "number"
          },
          "account": {
            "$ref": "#/components/schemas/ConnectAccountBody"
          },
          "field": {
            "$ref": "#/components/schemas/ConnectFieldBody"
          },
          "specializations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectSpecializationBody"
            }
          }
        },
        "required": [
          "account",
          "field",
          "specializations"
        ]
      },
      "UpdateSupervisorBody": {
        "type": "object",
        "properties": {
          "capacity": {
            "type": "number"
          },
          "field": {
            "$ref": "#/components/schemas/ConnectFieldBody"
          },
          "specializations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectSpecializationBody"
            }
          }
        }
      }
    }
  }
}